pyproject.toml
[project]
name = "bbcli"
version = "0.1.0"
description = "A fast, git-context-aware Bitbucket CLI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Version Control",
]
dependencies = [
    "click>=8.3.2",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
]

[project.scripts]
bb = "bbcli.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/bbcli"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
